refactor(test): name scoped-cache test files by subject + unify the scoped-cache radical
Two naming problems flagged in review:
- `*.coverage.test.ts` named files after the campaign that produced them, not the unit under
  test. Merge `items.coverage.test.ts` into the existing `items.test.ts` (one file per subject;
  reconciled the mock setups via vi.hoisted + a shared env), and rename the two that had no
  existing subject file: permissions.coverage → permissions.test.ts, controllers/items.coverage
  → controllers/items.test.ts.
- The scoped-cache tagging tests were named `*-cache-tags` while the rest of the feature uses the
  `scoped-cache` radical. Rename to match: items-/permissions-/graphql cache-tags →
  *-scoped-cache-tags (+ align two describe titles).

Renaming pulls each file's pre-existing long lines into the diff-scoped style gate (a rename
reads as a full add), so a handful of #203-era objects/titles got verticalized to fit 90 cols.
All tests still pass (items.test.ts now 65, coverage of items.ts unchanged).
